date picker moved to components in sections
